pyproject.toml
[project]
name = "scramble"
version = "0.1.0"
description = "TRS-80 Scramble arcade game - Python/SDL2 port"
requires-python = ">=3.10"
dependencies = [
    "pysdl2>=0.9.16",
    "pysdl2-dll>=2.28.5",
]

[dependency-groups]
dev = ["pytest>=8.0"]
